+++ jib-core/src/main/java/com/google/cloud/tools/jib/api/JibContainerBuilder.java
@@ -593,11 +593,6 @@ public class JibContainerBuilder {
         TimerEventDispatcher ignored =
             new TimerEventDispatcher(
                 buildContext.getEventHandlers(), containerizer.getDescription())) {
-
-      if (buildContext.getContainerConfiguration().getPlatforms().size() != 1) {
-        throw new UnsupportedOperationException(
-            "multi-platform image building is not yet supported");
-      }
       logSources(buildContext.getEventHandlers());
 
       BuildResult buildResult = containerizer.run(buildContext);