+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/json/V22ManifestListTemplate.java
@@ -128,6 +128,16 @@ public class V22ManifestListTemplate implements ManifestTemplate {
     public static class Platform implements JsonTemplate {
       @Nullable private String architecture;
       @Nullable private String os;
+
+      @Nullable
+      public String getArchitecture() {
+        return architecture;
+      }
+
+      @Nullable
+      public String getOs() {
+        return os;
+      }
     }
 
     @Nullable private String mediaType;
@@ -167,14 +177,13 @@ public class V22ManifestListTemplate implements ManifestTemplate {
      * @param os the manifest os
      */
     public void setPlatform(String architecture, String os) {
-      this.platform = new Platform();
-      this.platform.architecture = architecture;
-      this.platform.os = os;
+      platform = new Platform();
+      platform.architecture = architecture;
+      platform.os = os;
     }
 
-    @VisibleForTesting
     @Nullable
-    Platform getPlatform() {
+    public Platform getPlatform() {
       return platform;
     }
   }